refactor: redesign network layer with separated platform abstractions
- Extract client_connection from clientimpl with platform-specific implementations
- Add poller abstraction (epoll/kqueue/IOCP) for cross-platform I/O
- Add worker_pool, timer_queue, connection_table, runtime components
- Remove per-platform clientimpl files and connect_timeout
- Fix build.bat path handling for Windows packaging
- Add test_worker_pool, update test_integration and test_cpp_ownership
- Remove flaky test_log_seqlock

+++ example/client_cc.cc
@@ -11,6 +11,7 @@
 #include <connx/client.h>
 #include <connx/codec/delimiter_codec.h>
 #include <connx/options.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -66,8 +67,7 @@ public:
 
         connx::Metrics m;
         client_->GetMetrics(&m);
-        printf("[metrics] sent=%llu recv=%llu\n", (unsigned long long)m.bytes_sent,
-               (unsigned long long)m.bytes_received);
+        printf("[metrics] sent=%" PRIu64 " recv=%" PRIu64 "\n", m.bytes_sent, m.bytes_received);
         connx::ReleaseClient(client_); // deletes codec via ~ClientImpl
     }
